Split type impl generic method tests
