+++ docs/PHASE_PLAN.md
@@ -110,6 +110,8 @@ checked-in docs, tests, and commits only.
   from type, program, function, and global emission.
 - Resolver collection tests now keep behavior default-method synthesis cases in
   a focused behavior-methods submodule, separate from signature metadata cases.
+- Resolver collection tests now keep generic `Type.impl` method integrity cases
+  in a focused submodule, separate from template shape and mutability metadata.
 - Resolver-backed callable type-reference validation now shares the collected
   signature/body validation helper after function, top-level method, and
   `Type.impl` method dispatch restore the resolver-owned callable key, with
