+++ src/typechecker/tests/resolver_collection/function_method_templates/type_impl_generic_methods/integrity.rs
@@ -1,0 +1,169 @@
+use super::*;
+
+#[test]
+fn collect_declarations_with_symbols_does_not_fallback_to_stale_ast_generic_impl_method_template() {
+    let mut program = parse_program(
+        r#"
+Box: { value: i32 }
+
+Box.impl = {
+    keep<T> = (self: Box, value: T) T { value }
+}
+"#,
+    );
+    let mut symbols = crate::resolver::Resolver::new()
+        .resolve_program(&program)
+        .expect("resolver succeeds");
+    symbols.set_return_type_for_test(Namespace::Value, "Box.keep", None);
+    if let Declaration::ImplBlock { methods, .. } = &mut program.declarations[1] {
+        if let Declaration::Function {
+            params,
+            return_type,
+            ..
+        } = &mut methods[0]
+        {
+            params[1].ty = AstType::Named("Stale".to_string());
+            *return_type = Some(AstType::Named("AlsoStale".to_string()));
+        }
+    }
+    let mut tc = TypeChecker::new();
+
+    tc.collect_declarations_with_symbols(&program.declarations, &symbols);
+
+    assert!(
+            !tc.generic_methods.contains_key("Box.keep"),
+            "resolver-backed collection should not keep AST-only generic impl method templates when resolver signature metadata is incomplete"
+        );
+}
+
+#[test]
+fn collect_declarations_with_symbols_clears_stale_generic_impl_method_template_after_key_restore() {
+    let mut program = parse_program(
+        r#"
+Box: { value: i32 }
+
+Box.impl = {
+    keep<T> = (self: Box, value: T) T { value }
+}
+"#,
+    );
+    let mut symbols = crate::resolver::Resolver::new()
+        .resolve_program(&program)
+        .expect("resolver succeeds");
+    symbols.set_return_type_for_test(Namespace::Value, "Box.keep", None);
+    if let Declaration::ImplBlock {
+        type_name, methods, ..
+    } = &mut program.declarations[1]
+    {
+        *type_name = "Missing".to_string();
+        if let Declaration::Function {
+            name,
+            params,
+            return_type,
+            ..
+        } = &mut methods[0]
+        {
+            *name = "missing".to_string();
+            params[1].ty = AstType::Named("Stale".to_string());
+            *return_type = Some(AstType::Named("AlsoStale".to_string()));
+        }
+    }
+    let mut tc = TypeChecker::new();
+
+    tc.collect_declarations_with_symbols(&program.declarations, &symbols);
+
+    assert!(
+            !tc.generic_methods.contains_key("Missing.missing"),
+            "resolver-backed collection should clear the stale AST generic impl method template key after resolver key restoration"
+        );
+    assert!(
+            !tc.generic_methods.contains_key("Box.keep"),
+            "resolver-backed collection should clear the restored generic impl method template key when resolver signature metadata is incomplete"
+        );
+}
+
+#[test]
+fn collect_declarations_with_symbols_uses_resolver_type_impl_generic_method_template_target_and_name_metadata(
+) {
+    let mut program = parse_program(
+        r#"
+Box: { value: i32 }
+
+Box.impl = {
+    keep<T> = (self: Box, value: T) T { value }
+}
+"#,
+    );
+    let symbols = crate::resolver::Resolver::new()
+        .resolve_program(&program)
+        .expect("resolver succeeds");
+    if let Declaration::ImplBlock {
+        type_name, methods, ..
+    } = &mut program.declarations[1]
+    {
+        *type_name = "Missing".to_string();
+        if let Declaration::Function {
+            name,
+            params,
+            return_type,
+            ..
+        } = &mut methods[0]
+        {
+            *name = "missing".to_string();
+            params.pop();
+            *return_type = None;
+        }
+    }
+    let mut tc = TypeChecker::new();
+
+    tc.collect_declarations_with_symbols(&program.declarations, &symbols);
+
+    let template = tc
+        .generic_methods
+        .get("Box.keep")
+        .expect("generic impl method template");
+    assert!(!tc.generic_methods.contains_key("Missing.missing"));
+    assert_eq!(template.params.len(), 2);
+    assert_eq!(template.params[0].name, "self");
+    assert_eq!(template.params[1].name, "value");
+    assert_eq!(template.params[1].ty, AstType::Named("T".to_string()));
+    assert_eq!(template.return_type, Some(AstType::Named("T".to_string())));
+}
+
+#[test]
+fn collect_declarations_with_symbols_uses_resolver_type_impl_generic_method_name_for_body_type_refs(
+) {
+    let mut program = parse_program(
+        r#"
+Box: { value: i32 }
+
+Box.impl = {
+    keep<T> = (self: Box, value: T) T {
+        same: T = value
+        same
+    }
+}
+"#,
+    );
+    let symbols = crate::resolver::Resolver::new()
+        .resolve_program(&program)
+        .expect("resolver succeeds");
+    if let Declaration::ImplBlock { methods, .. } = &mut program.declarations[1] {
+        if let Declaration::Function {
+            name, type_params, ..
+        } = &mut methods[0]
+        {
+            *name = "missing".to_string();
+            type_params[0].name = "Stale".to_string();
+        }
+    }
+    let mut tc = TypeChecker::new();
+
+    tc.collect_declarations_with_symbols(&program.declarations, &symbols);
+
+    assert!(
+            tc.diagnostics.is_empty(),
+            "resolver-restored generic impl method name and type parameters should avoid stale AST body type-ref diagnostics: {:?}",
+            tc.diagnostics
+        );
+}
